+++ src/pages/validationNotificationTemplates/List.spec.tsx
@@ -1,13 +1,13 @@
-import schemaMock from '../../__mocks__/schema/constant.json'
+import schemaMock from '../../__mocks__/schema/validationNotificationTemplate.json'
 import { SECTIONS_MAP } from '../../lib'
-import { testConstant } from '../../testUtils/builders'
+import { testValidationNotificationTemplate } from '../../testUtils/builders'
 import { generateDefaultListTests } from '../defaultListTests'
 import { Component } from './List'
 
-const section = SECTIONS_MAP.constant
+const section = SECTIONS_MAP.validationNotificationTemplate
 const mockSchema = schemaMock
 const ComponentToTest = Component
-const generateRandomElement = testConstant
+const generateRandomElement = testValidationNotificationTemplate
 const customData = {}
 
 generateDefaultListTests({